perf:adjust queryConfig interface
