+++ arex-web-model-contract/src/main/java/com/arextest/web/model/contract/contracts/config/replay/ReplayConfiguration.java
@@ -8,11 +8,6 @@ import java.util.List;
 public class ReplayConfiguration {
 
     /**
-     * @see ScheduleConfiguration
-     */
-    ScheduleConfiguration scheduleConfiguration;
-
-    /**
      * comparison configuration
      */
     List<ReplayComparisonConfig> replayComparisonConfigs;
